+++ docs/API/changelog.md
@@ -14,6 +14,11 @@ tags:
 
 - Fixing ``plot_mode`` forwarding in radiometric analysis plots
 
+**Additional Features**
+
+- Radiometric Analysis: adding configuration for `River Masking` in Rain Forest analysis
+- Radiometric Analysis: adding Rain Forest full implementation function
+
 ## v2.1.2
 
 **Bug Fixing**